requirements.txt
matplotlib==3.5.2
numpy==1.21.5
pandas==1.4.4
PyYAML==6.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
